requirements.txt
google-generativeai
python-dotenv
pyperclip
requests